move AD glue to separate required files
