+++ src/reversediff.jl
@@ -2,23 +2,29 @@ module ReverseDiffX
 
 export NotTracked
 
-using MacroTools, LinearAlgebra, ..ReverseDiff, StaticArrays
+using LinearAlgebra
+using ..ReverseDiff
+using StaticArrays
+using ZygoteRules
+using Distributions
+using PDMats
+
 using Base.Broadcast: BroadcastStyle, ArrayStyle, Broadcasted, broadcasted
 using ForwardDiff: ForwardDiff, Dual
-using ..ReverseDiff: SpecialInstruction, value, value!, deriv, track, record!, tape, unseed!, @grad, TrackedReal, TrackedVector, TrackedMatrix, TrackedArray
+using ..ReverseDiff: SpecialInstruction, value, value!, deriv, track, record!,
+                     tape, unseed!, @grad, TrackedReal, TrackedVector,
+                     TrackedMatrix, TrackedArray
 using ..DistributionsAD: DistributionsAD
 
-const TrackedVecOrMat{V,D} = Union{TrackedVector{V,D},TrackedMatrix{V,D}}
 
 import SpecialFunctions, NaNMath
 import ..DistributionsAD: turing_chol, symm_turing_chol, _mv_categorical_logpdf
 import Base.Broadcast: materialize
 import StatsFuns: logsumexp
-import ZygoteRules
 
+const TrackedVecOrMat{V,D} = Union{TrackedVector{V,D},TrackedMatrix{V,D}}
 const RDBroadcasted{F, T} = Broadcasted{<:Any, <:Any, F, T}
 
-using Distributions, PDMats
 import Distributions: logpdf,
                       Gamma,
                       MvNormal,
@@ -29,15 +35,15 @@ import Distributions: logpdf,
                       PoissonBinomial,
                       isprobvec
 
-using ..DistributionsAD: TuringMvNormal,
-                         TuringMvLogNormal,
-                         TuringWishart,
-                         TuringInverseWishart,
-                         TuringDirichlet,
-                         TuringPoissonBinomial,
-                         TuringScalMvNormal,
-                         TuringDiagMvNormal,
-                         TuringDenseMvNormal
+using ..DistributionsAD: TuringPoissonBinomial
+#                                      TuringMvNormal,
+#                                      TuringMvLogNormal,
+#                                      TuringWishart,
+#                                      TuringInverseWishart,
+#                                      TuringDirichlet,
+#                                      TuringScalMvNormal,
+#                                      TuringDiagMvNormal,
+#                                      TuringDenseMvNormal
 
 include("reversediffx.jl")
 
@@ -71,6 +77,7 @@ function Base.maximum(d::LocationScale{T}) where {T <: TrackedReal}
     end
 end
 
+#=
 for T in (:TrackedVector, :TrackedMatrix)
     @eval begin
         function logpdf(d::MvNormal{<:Any, <:PDMats.ScalMat}, x::$T)
@@ -295,5 +302,6 @@ _mv_categorical_logpdf(ps::TrackedMatrix, x) = track(_mv_categorical_logpdf, ps,
         return (ps_grad, nothing)
     end
 end
+=#
 
 end
